Bump deps and figwheelize. Put system attrs at db 1
newer datascript doesn't like a datom with an id of 0 so move the
system attributes to id 1

Also, a curious bug about the filter text. Possibly a change in the
regex handling of clojurescript?

This diff was always returning [""] and so was trying to match things
on the empty string which didn't match (strange?). But it seems that
`(str/split nil #"\s+")` returned nil in 1.7.xxx and now returns [""]
in 1.10.520.

(not-empty
-    (str/split (system-attr db :system/filter) #"\s+")))
+    (remove str/blank?
+            (str/split (system-attr db :system/filter) #"\s+"))))